Nemelex changes: - Renamed abilities, and reordered so that you get Draw One before Peek Two - Mark Four and Peek Two now lose some cards from the deck - Viewing a deck now shows which cards have been marked in it Also fixed a small warning in format.cc.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@4059 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/format.cc
@@ -351,7 +351,7 @@ std::string formatted_string::to_colour_string() const
             
             while (true)
             {
-                const unsigned int left_angle = st.find('<', start);
+                const size_t left_angle = st.find('<', start);
                 if (left_angle == std::string::npos)
                     break;
